refactor: reorganize visualization components
