+++ src/components/analytical/CorrelationDetailChart.tsx
