+++ src/components/analytical/CorrelationDetails.tsx
