+++ src/components/analytical/CorrelationRippleMatrix.tsx
