+++ src/components/analytical/PaceDeltaHistogram.tsx
