+++ src/components/analytical/SessionDetailDrawer.tsx
