+++ src/components/analytical/__tests__/CorrelationRippleMatrix.test.tsx
