+++ src/components/analytical/__tests__/SessionDetailDrawer.test.tsx
