+++ src/components/examples/GhostSelfRivalChart.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import WeeklyComparisonChart from '@/components/statistics/WeeklyComparisonChart'
+import WeeklyComparisonChart from '@/components/trends/WeeklyComparisonChart'
 
 export const description = 'Compare this week with the same week last year'
