+++ src/components/goals/GoodDayBadges.tsx
