+++ src/components/goals/index.ts
@@ -1,0 +1,3 @@
+export { default as GoodDayBadges } from './GoodDayBadges'
+export { default as GoodDayForecastCalendar } from './GoodDayForecastCalendar'
+export { default as GoodDayInsights } from './GoodDayInsights'
