+++ src/components/maps/GeoActivityExplorer.tsx
