+++ src/components/maps/LocationEfficiencyComparison.tsx
