+++ src/components/maps/SessionSimilarityMap.tsx
