+++ src/components/maps/__tests__/GeoActivityExplorer.test.tsx
