--- src/components/statistics/index.ts
@@ -1,16 +1,0 @@
-export { default as ActivityByTime } from "./ActivityByTime";
-export { default as AvgDailyMileageRadar } from "./AvgDailyMileageRadar";
-export { default as GoodDayMap } from "./GoodDayMap";
-export { default as GoodDaySparkline } from "./GoodDaySparkline";
-export { default as GoodDayInsights } from "./GoodDayInsights";
-export { default as GoodDayTrendline } from "./GoodDayTrendline";
-export { default as GoodDayBadges } from "./GoodDayBadges";
-export { default as GoodDayForecastCalendar } from "./GoodDayForecastCalendar";
-export { default as HabitConsistencyHeatmap } from "./HabitConsistencyHeatmap";
-export { default as PeerBenchmarkBands } from "./PeerBenchmarkBands";
-export { default as PerfVsEnvironmentMatrix } from "./PerfVsEnvironmentMatrix";
-export { default as SessionSimilarityMap } from "./SessionSimilarityMap";
-export { default as TreadmillVsOutdoor } from "./TreadmillVsOutdoor";
-export { default as WeeklyComparisonChart } from "./WeeklyComparisonChart";
-export { default as FilterBar } from "./FilterBar";
-
