+++ src/components/trends/FocusTimeline.tsx
