+++ src/components/trends/GoodDaySparkline.tsx
