+++ src/components/trends/GoodDayTrendline.tsx
