+++ src/components/trends/HabitConsistencyHeatmap.tsx
