+++ src/components/trends/WeeklyComparisonChart.tsx
