+++ src/components/trends/__tests__/WeeklyComparisonChart.test.tsx
