+++ src/components/trends/index.ts
@@ -1,0 +1,6 @@
+export { default as FocusTimeline } from './FocusTimeline'
+export { default as ActivityByTime } from './ActivityByTime'
+export { default as GoodDayTrendline } from './GoodDayTrendline'
+export { default as GoodDaySparkline } from './GoodDaySparkline'
+export { default as HabitConsistencyHeatmap } from './HabitConsistencyHeatmap'
+export { default as WeeklyComparisonChart } from './WeeklyComparisonChart'
