--- src/components/visualizations/index.ts
@@ -1,6 +1,0 @@
-export { default as BehavioralCharterMap } from "./BehavioralCharterMap";
-export type { Segment } from "./BehavioralCharterMap";
-export { default as TransitionMatrix } from "./TransitionMatrix";
-export { default as CorrelationRippleMatrix } from "./CorrelationRippleMatrix";
-export { default as CorrelationDetails } from "./CorrelationDetails";
-export { default as FocusTimeline } from "./FocusTimeline";
